pyproject.toml
[tool.poetry]
name = "solar-works-job-nimbus"
version = "2.0.0"
description = ""
authors = ["HM <hieumdd@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
selenium = "^3.141.0"
selenium-wire = "^4.5.2"
requests = "^2.26.0"
google-cloud-bigquery = "^2.26.0"
functions-framework = "^2.2.1"
google-cloud-tasks = "2.5.1"
compose = "^1.2.8"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest = "^6.2.5"
mypy = "^0.910"
types-requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-s --durations=0"
testpaths = [
    "test",
]

[tool.mypy]
python_version = 3.9
ignore_missing_imports = true
namespace_packages = true
